shipment-api/requirements.txt
databases[asyncpg]==0.9.0
dependency-injector==4.42.0
fastapi==0.115.4
metar==1.11.0
pydantic==2.10.1
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
uvicorn==0.32.0
passlib~=1.7.4
asyncpg~=0.30.0
python-jose==3.3.0
python-multipart==0.0.17
geopy~=2.4.1
haversine==2.9.0
fastapi-mail==1.4.2
httpx==0.28.1
pytest==8.3.5
coverage==7.8.0